backend/mart/pyproject.toml
[tool.poetry]
name = "mart"
version = "0.1.0"
description = ""
authors = ["Uzma <nasiruddin.uzma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
stripe = "^10.11.0"
sqlmodel = "^0.0.22"
requests = "^2.32.3"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
pydantic = "^2.9.2"
aiokafka = "^0.11.0"
celery = "^5.4.0"
passlib = "^1.7.4"
pytest = "^8.3.3"
python-json-logger = "^2.0.7"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
pyjwt = "^2.9.0"
python-multipart = "^0.0.12"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
requests = "^2.32.3"
coverage = "^7.6.1"
pytest = "^8.3.3"

[tool.black]
# https://github.com/psf/black
line-length = 88
target_version = ['py38']
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''

[tool.isort]
# https://github.com/timothycrosley/isort
# https://github.com/timothycrosley/isort/wiki/isort-Settings
line_length = 88
indent = '    '
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0


[tool.coverage]
# https://github.com/nedbat/coveragepy
    [tool.coverage.run]

    source = ["app"]
    branch = true
    omit = ['']

    [tool.coverage.report]
    exclude_lines = [
        "# noqa",
        "raise NotImplementedError"
    ]



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
